pyproject.toml
[tool.poetry]
name = "laptoplens"
version = "0.1.0"
description = ""
authors = ["hamine95 <hamine95>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
celery = "^5.4.0"
python-decouple = "^3.8"
django-celery-beat = "^2.6.0"
django-celery-results = "^2.5.1"
redis = "^5.0.4"
ollama = "^0.1.9"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
